requirements.txt
opencv_python==4.1.1.26
pandas==1.0.3
matplotlib==3.1.3
numpy==1.18.1
Pillow==8.2.0
scikit_learn==0.24.2
